Refactor to move the isTick to a utility class
