+++ storm-core/src/jvm/backtype/storm/utils/TupleUtils.java
@@ -15,27 +15,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package storm.starter.tools;
+package backtype.storm.utils;
 
 import backtype.storm.Constants;
 import backtype.storm.tuple.Tuple;
 
-import static org.mockito.Mockito.*;
+public final class TupleUtils {
 
-public final class MockTupleHelpers {
-
-  private MockTupleHelpers() {
+  private TupleUtils() {
+    // No instantiation
   }
 
-  public static Tuple mockTickTuple() {
-    return mockTuple(Constants.SYSTEM_COMPONENT_ID, Constants.SYSTEM_TICK_STREAM_ID);
-  }
-
-  public static Tuple mockTuple(String componentId, String streamId) {
-    Tuple tuple = mock(Tuple.class);
-    when(tuple.getSourceComponent()).thenReturn(componentId);
-    when(tuple.getSourceStreamId()).thenReturn(streamId);
-    return tuple;
+  public static boolean isTick(Tuple tuple) {
+    return tuple != null
+           && Constants.SYSTEM_COMPONENT_ID  .equals(tuple.getSourceComponent())
+           && Constants.SYSTEM_TICK_STREAM_ID.equals(tuple.getSourceStreamId());
   }
 
 }
